fix: error for Project Service

+++ module-project/src/main/java/com/gbc/codingmates/service/project/ProjectService.java
@@ -5,14 +5,11 @@ import com.gbc.codingmates.domain.project.ProjectRepository;
 import com.gbc.codingmates.dto.member.MemberDto;
 import com.gbc.codingmates.dto.project.ProjectDto;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.client.HttpClientErrorException;
-
-import javax.persistence.EntityManager;
-import javax.validation.Valid;
+import java.nio.file.AccessDeniedException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,40 +20,49 @@ import java.util.stream.Collectors;
 public class ProjectService {
 
     private final ProjectRepository projectRepository;
+    private final ModelMapper modelMapper;
 
 //    Create project
     @Transactional
-    public ResponseEntity<Long> saveProject(final ProjectDto ProjectDto){
-        validateDuplicateProject(ProjectDto);
-        Project project = projectRepository.save(ProjectDto.toEntity());
+    public ResponseEntity<Long> saveProject(final ProjectDto projectDto){
+        validateDuplicateProject(projectDto);
+        Project project = projectRepository.save(modelMapper.map(projectDto,Project.class));
 //        return new ResponseEntity<>("Hello World!", HttpStatus.OK);
         return ResponseEntity.ok(project.getId());
     }
 
     //list all projects
-    public List<Project> listAll(){
+//    public List<Project> listAll(){
+////        return projectRepository.findAll();
+////        List<Project> projects = em.createQuery(
+////                "select p from Project p" +
+////                        " join fetch p.member m", Project.class
+////        ).getResultList();
+////        return projects;
+////        return projectRepository.listAllWithMember();
 //        return projectRepository.findAll();
-//        List<Project> projects = em.createQuery(
-//                "select p from Project p" +
-//                        " join fetch p.member m", Project.class
-//        ).getResultList();
-//        return projects;
-        return projectRepository.listAllWithMember();
+//    }
+
+    public ResponseEntity<List<ProjectDto>> listAll(){
+        List<Project> projectList = projectRepository.findAll();
+        return ResponseEntity.ok(projectList.stream()
+                .map(project -> modelMapper.map(project,ProjectDto.class))
+                .collect(Collectors.toList()));
     }
+
 
     public ResponseEntity<ProjectDto> findById(final ProjectDto projectDto){
         Project project = projectRepository.findById(projectDto.getId()).orElseThrow(() -> new IllegalArgumentException());
-        return ResponseEntity.ok(projectDto);
+        return ResponseEntity.ok(modelMapper.map(project,ProjectDto.class));
     }
 
     //update/edit project
     @Transactional
-    public ResponseEntity<Long> edit(final MemberDto memberDto, final Long id, final ProjectDto ProjectDto){
+    public ResponseEntity<Long> edit(final MemberDto memberDto, final Long id, final ProjectDto ProjectDto) throws AccessDeniedException {
         Project project = projectRepository.findById(id).orElseThrow(() -> new IllegalArgumentException());
-        Long ownerId = project.getId();
+        Long ownerId = project.getMember_id();
         if(ownerId!=memberDto.getMemberId()){
-//            throw HttpClientErrorException.Forbidden;
-//            return ResponseEntity.badRequest(memberDto.getMemberId());
+            throw new AccessDeniedException("you are not the owner of this post");
         }
         project.update(ProjectDto.getTitle(), ProjectDto.getContent());
         return ResponseEntity.ok(id);
@@ -64,8 +70,12 @@ public class ProjectService {
 
     //delete project
     @Transactional
-    public ResponseEntity<Long> deleteById(final Long id){
+    public ResponseEntity<Long> deleteById(final MemberDto memberDto, final Long id) throws AccessDeniedException {
         Project project = projectRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("no such project"));
+        Long ownerId = project.getMember_id();
+        if(ownerId!=memberDto.getMemberId()){
+            throw new AccessDeniedException("you are not the owner of this post");
+        }
         projectRepository.delete(project);
         return ResponseEntity.ok(id);
     }
